Double instead of Int?
If I were you I would do use Double instead of Int, because the most times when you make a plugin and will use the getBlock(); method. You  will love it to add 0.5 to it. Otherwise players will teleport in to the ground (and setting that 1 block higher also doesn't look very nice).