Create Location class
This is a (basic - needs some work) general world-related class that encapsulates the world, x, y, and z coordinates, and the float and yaw of a point. 
I was a bit hesitant to submit this since I notice there was (and I believe still is) some ongoing discussion about vectors, but I'll submit this nonetheless.
Note: many of the methods still require documentation.