Add PlayerConnection. PlayerConnection represents the active connection of a client player to the server.
Add ChannelBuf. ChannelBuf is an abstract interface
to represent a ByteBuffer to allow plugins and the server
to read and write data without depending on a specific
networking library.

Add CustomChannelListener for listening to
custom payloads. This should facilitate some simple handling of
payloads by plugins.

Add ChannelRegistrationException for when channels are already registerd.