Remove AreaBuffer and VolumeBuffer
New methods to retrieve the size/location have been added to the
BiomeBuffer and BlockBuffer interface. It is now possible to create a
class that implements both BiomeBuffer and BlockBuffer.

Also fixed a typo in BiomeArea, and added missing getBlock(Vector3i) to
BlockBuffer and added some this. prefixes to PopulatorObjects.