Fixes for the custom world generator API
GameRegistry changes:
Now that the id of a world generator modifiers is supplied by the
modifier itself, the id is no longer needed in the register method.

Also added some requirements to the id returned by
WorldGeneratorModifier.

Give world generator modifiers more context:
With the removal of the
generator settings from WorldGenerator, the
WorldGeneratorModifier
interface was damaged: there was no way to obtain access to the world
seed anymore.

This commit fixes that by replacing the worldname
parameter with the
newly introduced WorldCreationSettings interface.