Add TileEntityChangeEvent.
In the result of an event having the concept of changing a TileEntities'
data manipulator (besides doing so in TileEntity.offer, shame on you),
the event in question should extend this event and use getNewData after
resolution.

The concept at play here is preventing needless sets to the data until the
event is done and ultimately the last modifier of the data sticks.

Signed-off-by: Chris Sanders <zidane@outlook.com>