Move Extent#contains(Location) to Location
The problem was that an extent might have multiple different data
bounds, for example biomes and blocks. There is no guarantee that these
have the same size. This made the method unclear: what must the extent
contain from the location for it to be contained? The block, the biome,
both? In other words, if the location has a biome in the extent, but
not a block, is it contained in the extent?

Instead, this adds methods to Location to check if it has a biome or a
block at its position. A method was also added to check if the location
is in a given extent. This solves the issue described above by
separating the different data types.