Start Value API additions for Data API 2.0.
Values as of current writing are proxies to their owners, being
able to get and set a specific value of information that is
keyed to a particular Key. ValueStores are able to store multiple
values and their Value(s) can be retrieved through Keys. When
working with Values, live editing can be performed, without the
worry of having to offer data back. For the most part, the
DataManipulators are no longer manipulating data directly, but
are containers of Values. DataHolders now work with values
directly as well as DataManipulators.