Complete Value API integration with Data API. Completes Data API 2.0.
This unifies the original intention for Values with the intentions of
Data API 1.0. All forms of changes of data are transactional, with the
exception of DataManipulators. As DataManipulators remain simply
containers of some values, mutating them are non-transactional.

With regards to Values, a Value does not contain a link to it's owner,
but rather it is a wrapper surrounding an object that it holds. From
a Value, we can gather that the data it wraps is "keyed" for a
particular use such that two Values can have the same data, but can
have different Key(s) from one and the other. Values themselves have
been reduced to intermediary objects for handling data from
ValueContainers and DataHolders.

The hierarchy of Data is now as follows:

BaseValue
 - Value
  - OptionalValue
  - MutableBoundedValue
  - MapValue
  - CollectionValue
   - SetValue
   - ListValue
   - WeightedCollectionValue
 - ImmutableValue
  - ImmutableOptionalValue
  - ImmutableBoundedValue
  - ImmutableMapValue
  - ImmutableCollectionValue
   - ImmutableSetValue
   - ImmutableListValue
   - ImmutableWeightedCollectionValue
 - BoundedValue
  - MutableBoundedValue
  - ImmutableBoundedValue

ValueContainer
 - DataManipulator
 - ImmutableDataManipulator
 - CompositeValueStore
  - DataHolder
 - ImmutableValueStore
  - ImmutableDataHolder
   - BlockState
   - PotionEffect

From the perspective of all these changes, much has been retained
with regards to how data was previously manipulated. The largest
difference is the now split trees of mutable and immutable versions
of DataManipulators, Values, and DataHolders. The benefit of these
changes result in the ability to retrieve data from a DataHolder,
and translate it into an immutable version of itself. The advantage
clearly now is that any events providing a DataManipulator, can now
provide an ImmutableDataManipulator in its stead, without having to
perform copies and or use other protective measures. Of course,
having the immutable versioning of both DataManipulators and Values
allows for DataTransactionResults to become truly transaction
results. Furthermore, immutable versions can be cached for repeated
use, so long as the values they contain are immutable.

While the inclusion of Values allows for more precise data
manipulation, Keys are able to identify data for more than just
data manipulation. A Key, while being an actual key to a piece of
information potentially stored on a ValueContainer, the key can
also dictate the DataQuery associated with the piece of information
and be relied upon for serialization and deserialization purposes.

Lastly, additions to Properties for Locations have been made, and
registration of PropertyStores for custom properties is now available.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>