Refactor Location methods to be more specific
When this class was only BlockLoc, adding the word “block” made the
names redundant as the class only dealt with blocks. Now that this is a
generic location class most methods have confusing names, such as
getSnapshot(), replaceWith(), getType(), remove(), etc. The names
have been changed to match the ones in Extent.