Graduate UUIDs to Extent, change data returned in Location#toContainer
This commits is dual purposed. First off, all Extents should have the
ability to determine unique-ness (a feature that was solely for World
before).

Second, BlockType and Manipulators are an unknown factor at any
given time when working with Locations (proven by the fact that game logic
location methods proxy back to the Extent). As such, toContainer should
never return "maybe" data; this commit addresses this.

Signed-off-by: Chris Sanders <zidane@outlook.com>