Add API to allow inserting placeholders in a Text
This API allows inserting placeholder objects (Texts.placeholder())
into a text object, which can later be easily replaced (with the
Texts.format methods).

Additionally, this allows click, hover, and shift click actions to be
specificed in Texts.of() as necessary.

The interface TextRepresentable is added to allow more complicated
objects to define their own Text representation, and appropriate Texts
methods are updated as necessary.

Closes #779. Closes #639.