Add API for Text onClick callbacks
This API assumes that every text can only be interacted with by a
CommandSource, and passes said CommandSource to the callback. This
allows easily sending one Text instance to multiple CommandSources, but
would cause issues if someone came along trying to have a
non-CommandSource interact with Text. However, I think that is a pretty
unlikely occurrence.

This also adds a Consumer functional interface since we don't have
access to JDK8 interfaces (would be pretty useful :(), which is used as
the interface callbacks implement.

Closes #719.