Add PlayerLogin and PreLogin events
This additionally required minor changes to the event factory test to
properly mock UUID instances -- because UUID is final Mockito cannot
give us instance, so another case had to be added to generate a random
UUID for this.

Furthermore, these events are incomplete without a full Cause
implementation, meaning that some information about *why* a player
might be kicked is not yet available. In the future, the cause methods
would return more useful objects. For example:

- If a player were banned, a Ban object would be provided with
  information about the player's ban

- If a player were not whitelisted on a server, the server Whitelist
  would be the cause -- this works differently because whitelist kicks
  are done based on an absense of an entry, as opposed to the bans
  blacklist.

- If a player is unable to join due to bypassing the server player
  count limit, some wrapper around the maximum player count for the
  server would be returned.

Keep in mind these examples are merely suggestions, and may change
significantly in the process of implementation. Keep in mind that only
the login event is cancellable -- the other two in the join trio
(prelogin, join) cannot be cancelled. This is to keep a consistent
sequence of operations that can be done at different phases, meaning a
plugin will only have to listen to one event to learn about which
players are prevented from joining.

The pipeline for joining would look like this (from an event perspective)

1. PlayerPreLoginEvent: called asynchronously, very early in the login
process. Should be used to cache any potentially long-running data
lookups for player information.

2. PlayerLoginEvent: called synchronously after a player object has
been created but before final join information has been determined. Used
to determine whether a player will in fact be allowed to join the server.
This is the only cancellable event in the squence (keep in mind that
cancelling this event *will not* throw any sort of quit or kick events, so
listen to this on the latest priority possible to catch potential
cache removals).

3. PlayerJoinEvent: called synchronously as a player is being spawned
into the world. This event allows changing player join state -- spawn
location, join message, etc, but is not cancellable.