Add EntityInteractionType.INDIRECT
As the javadocs elaborate on, this is meant for "indirect" interactions
where the server knows that an entity has performed an interaction but it
wasn't a direct one by that entity.

Examples include trampling crops, walking over tripwire, walking on top of
pressure plates, etc...anything where the player wasn't performing the
interaction directly.

Signed-off-by: Chris Sanders <zidane@outlook.com>