Improve PlayerChatEvent for better plugin interop
This change makes use of placeholders to have an api-exposed method of
interoperability for chat messages. The unformatted message is also
made mutable so that plugins can change the player's message without
having to be concerned with formatting.

These changes have been made to make it easier for multiple plugins to
have a say in the final message sent by a player.