Allow BlockSnapshots to be restored without location.
Since a BlockSnapshot contains the position of the BlockState it captured,
it should be allowed to be restored using the recorded position.

* Added restoreSnapshot with new parameters force and notifyNeighbors.

The parameter 'force', if true, will restore the BlockSnapshot state even if the BlockType does not match.

The parameter 'notifyNeighbors' ,if true, will restore the BlockSnapshot then notify all neighboring blocks triggering physic updates.

* Updated javadocs