Replace optional TextArg return
An empty state already exists in the form of Texts.of(). Adds an
isEmpty() method to Text to check if an instance is empty. This means
that the empty Text can be used instead of an absent optional.
TextArgs like iterator() or iterable() are handled accordingly, not
appending the result if it is empty.