Add back setters to Transform but it returns new variants.
While we do have a Transform builder, its a lot simpler to just be able to
use an existing one and set the attribute you desire and get a new copy.

The transform is still desired for creating a transform from scratch or
doing mass changes to an existing transform.

Signed-off-by: Chris Sanders <zidane@outlook.com>