Remove Attributes API
The current Attributes API has major issues, which prevent it from being
implemented at all.

Because of this, the current API is being completely removed. Once
significant changes have been made to allow it to be implemented, it
will be re-added.

Plugins, documenters, and any other interested parties should treat
the Attributes API as non-existent, until it is re-added in an improved
form.