Expose more of the logging interfaces
This API functions as a layer on top of slf4j, both providing increased
access to logging appenders and filters, as well as allowing the
logging of Text messages rather than just plain Strings.

These changes allow for loggers to have improved integration with
Sponge and for plugins to more easily intercept plugin log messages.

Additionally, the concept of an action logger is added. This is more
complicated than the existing API changes, and is designed for plugins
that perform actions that have a source, one or more targets, and the
need to notify some combination of those and the log/console of events
that occur.