Add support for extended BlockStates.
Extended BlockStates include extra properties about a given location such
as if a podzul dirt block is covered by snow or a fence is connected to
other fences.

There are currently 2 ways to retrieve an extended BlockState :

1. From a BlockSnapshot by calling getExtendedState.
2. From a BlockState by calling withExtendedProperties. This method should
only be called on live BlockStates and may not work correctly if used
with a BlockSnapshot such as during a ChangeBlockEvent.