Phase 1 AI Framework - Goals. Documentation soon to follow.
Rename Human to Humanoid, Make Human both Humanoid and Creature.

With what may seem like an odd change is necessary because we desire to
share many things between Player and Human.

Humans in Sponge are no different than a Skeleton, Zombie, etc...they are
just "Creature"s. As a Creature, I can easily give them AI in a plugin
which means the obvious choice would be Human IS-A Creature. This presents
problems because Player would then be capable of AI and while that seems
silly...if the AI allows it, someone will attempt it and this would fail
epically.

Simple fix is a common interface between Human and Player and keep AI to
Human for Humanoid types.

Lastly, add the ability to remove tasks.

Add BoundAITasks/AIBuilders for when a task has a known Owner.

BoundAITasks are tasks in-which we, for sure, know the owner because the
task is being created. This is a difference from AbstractAITask where, due
to it being wrapped in the impl, we don't know the owner until it being
added.

AIBuilders are used to allow the API to create Vanilla AI objects to
attach to entities. Currently, I've provided a builder for the WanderAI so
that we can see how this will work.

Need to return a List of AITasks.

Its entirely possible to have more than one instance of an AITask on a
single entity. We support this mechanic as well.

Make the owner of the AITask a core concept.

Finish up most of core AI structure.

I've brought back Goals as each Minecraft entity has one Goal with
sometimes having two (namely Monsters).

Updated javadocs on everything.

Some feedback from @Aaron1011.

Place generic on type to enforce task, type, and goal all have the same
agent type.

Some doc corrections.

Reverse requiring Owner on type.

While it seems to be the most sensible solution as it allows us to do
compile-time verification that the owner class on the type matches the
AbstractAITask's owner, the problem it presents is the fact that getting
it from the GameRegistry becomes a double-casting type erasure funfest.
The changes made fixes this at the cost of runtime verification.

Also, due to fullwall's feedback, I've removed mutexes. Now, there is a
method to determine conconcurrency.

Signed-off-by: Chris Sanders <zidane@outlook.com>