Add abstract implementation of DataManipulator and ImmutableDataManipulator.
This allows for plugins to easily rely on these abstractions without having
to ipmlement all of the considered boilerplate code with Keys and Values.
It is intended that these abstractions do not provide any implementation of
creating mutable instances from a reflection util as seen in the Common
implementation of SpongeAPI. As well, caching is not provided so as to
prevent abuse of a potentially delicate and misused system.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>