Separate value CommandElements from structural ones
This improves type-safety. It also opens the door for usability
improvements in CommandContext and downstream command implementations.

Additionally, this removes some awkward cases where structural elements
had to implement a bunch of extra stuff.

However, behavior will have to change in some places, so that should be
kept in mind.