Level 3: Make CatalogType and World a TextRepresentable
CatalogTypes are often used in commands and they have a discrepancy
between their human readable name and their internal id which requires
them to use special representations to be usable in commands.
Worlds are often used in chat messages and they have a name so, the name
should be displayed.