Stop providing the populated transactions in Post.
It wasn't the goal of events to ever replace what the World Generation API
does and we also don't need to be generating thousands upon thousands of
snapshots.

With that said, I understand the desire to have some sort of callback when
population occurs so a compromise is in place and the chunk + populators
applied are left available.

Signed-off-by: Chris Sanders <zidane@outlook.com>