Remove world (folder) name from WorldCreationSettings and catalog it.
WorldCreationSettings was never meant to represent instance World data: it
was always meant to be used to build World templates and used to create
many worlds. This commit enforces this concept by letting the plugin
developer create these templates and use them for creating any number of
worlds.

As a consequence, establishing the world's folder name is now specified in
Server#createWorldProperties.

Finally, SpongeAPI ships the default settings types for the Vanilla-style
Worlds found in Minecraft.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>