Change 'ChangeBlockEvent.Pre' to support bulk locations.
In order to support plugins that do bulk editing such as WorldEdit, the
Pre event must be able to support one or more locations. The idea is for
the plugin to pass their PluginContainer as root cause and if the event is
not cancelled, no tracking or events will occur.