Add required dependency on SpongeAPI for plugins
The dependency version is set to the SpongeAPI version the plugin is
compiled against. Without additional restrictions the dependency
version will just represent the "recommended version" for the
dependency.

While plugins will continue to load using newer or older API versions,
these changes allow the implementation to issue a warning if a plugin
was designed for an older or newer API version.