Modify child command logic to allow for meaningful usage
One of the long standing issues with the Command API is that subcommands don't give useful feedback if a command fails. The current implementation of the command system is difficult to solve this effectively with - this is a best effort. It isn't breaking, but in order to activate behaviour where a subcommand failing gives meaningful usage requires "childArgumentParseExceptionFallback" on the CommandSpec builder to be false. It defaults to true, the current behaviour.

Any plugin that doesn't use this method in the builder will get the old behaviour.

Fixes #1165, fixes #1272, fixes #1542, fixes #1647