Gradle: Apply IntelliJ IDEA inheritOutputDirs fix to all modules
We need to set inheritOutputDirs to true for all IntelliJ IDEA
modules to avoid problems caused by separate output directories
for classes/resources in IntelliJ IDEA.

Up to now it was applied separately for SpongeAPI and the
implementation modules, which means it wasn't applied to the
testplugins module automatically. Move it to the shared Gradle
configuration to ensure it gets applied for all modules.