Change SkinData to use a custom ProfileProperty
It's often useful to be able to set an arbitrary signed
skin payload for a Humanoid:

* Plugins may want to use a skin that is not currently set on any
player
* The Mojang skin API has notoriously strict IP-based rate-limiting.
Servers, especially those that are part of a larger network, may
wish to implement their own caching system.

This commit allows plugins to set use valid skin ProfileProperty
with SkinData, rather than being constrained to use a skin currently
in use by a specific player.