Modify API to support Sponge Schematic Format Specification version 2.
Specifically, this adds non-binary-breaking changes to Schematic and
ArchetypeVolume and their related builders so that previous usages of
the API will be able to support the new Schematic format out of the box.
The default implementations are only provided for external
implementations of the Schematic API, so they do not break when plugin
authors use them. By default, new ArchetypeVolumes being created may or
may not include EntityArchetypes, but the implementation will support
them. Likewise, biome support is provided through Extents when creating
a Schematic, as ArchetypeVolume is not suitable for biome support.

Future API versions may break this API to add better capabilities with
ArchetypeVolumes and Schematics.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>