Data API improvements. DataStore and DataProviders
This commit highlights a change to allow the usage of Keys not
specifically through serialization/deserialization from ValueContainers,
but also by proxy through DataProviders. The advantage here of having
DataProviders and DataStores is that plugins can take advantage of using
the Data API convenience methods and still expose their own custom data
through DataProviders, which "provide Values indirectly of the
ValueContainer."

DataStores are the serializers for the Values given an appropriate
serialization targets based on the TypeToken of which they are
registered to. Examples for this differentiation include: DyeColor being
serialized for Sheep versus ItemStacks of Armor versus Shulker
BlockEntities. All store the same Value differently, but they're still
accessible the same way. The advantage of this is that by eliminating
the definition of knowing what DataQueries are used, Keys no longer
expose what DataQuery they are paired to, and Sponge no longer provides
a different format for vanilla objects. This means that raw data and
toContainer() methods will become virtual identical duplicates with the
exception of ContentVersion.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>