Value.asMutableCopy, add remove operations in DataProvider, rename ValueContainer view methods.
asMutableCopy is helpful when you don't know what value type you have and you want to be sure to create a copy. You can also do this with value.asMutable().copy(), but if the target value is immutable, the copy will be made twice, which is something that should be avoided.