Remove loading a world by it's UUID.
While I'm sure someone will question this change, I will explain why:

The problem comes down to the fact that Worlds (now and for the forseeable future)
are very much IO based. Keeping this method forces the impl to do one of three things:

A. Fail to load as it doesn't know of the UUID as it doesn't...
B. Cache all UUIDs that get loaded. If one is unloaded, remember that entry. Load it again later
   if it isn't know then...
C. Actually scan the save container for all UUIDs (!).

It just isn't worth pretending that the World management is anything different than that.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>