Rework the command API for API 8 (1.14)
Much of this is geared towards removal of CommandSources in the API and as such, some specialised types have been removed. However, there are a couple of unresolved questions:

* How do we represent an RCON client and the console - and where? The console is currently directly the Server object at this point. This will likely happen with the Client object too.
* The cause is an important part of our command system for 1.14+ and command writes should embrace it. However, we should consider a way to make it easy for developers to take our cause and work with our recommendations. One thing I'm thinking about is creating a replacement for the command source that wraps the cause with a few helper methods - much like the command context, but for commands that don't use our builder. This would be important when the cause event context has a part to play as that is intended to potentially override what's in the direct cause. That way, we can try to standardise what happens.

The console source has been removed in favour of using the `Server` as a command source. Signs are now directly sources - the fact the source is separate is, in my opinion, an implementation detail. Again, same with command blocks.

I have created javadocs accordingly, but this is up for discussion within the Sponge team and we can go either way at this point. There are implementation concerns too, but that's not the concern of the API!

This is all up for discussion. Things will change as implementation progresses.