Data API improvements. DataManipulator and DataHolder Simplification.
This brings along cleanup changes to DataHolder, ValueContainer,
DataManipulator, and their usages. Specifically, simplified the class
hierarchy of DataManipulator such that Mutable and Immutable types share
similar methods for #asMutable() and #asImmutable() along with proper
overrides for #copy(). The same treatment has been applied to
DataHolder, where DataHolder now superclasses it's mutable and immutable
counterparts, providing the extension of PropertyHolder and
DataSerializable.

Added Has, Supports, and GetValue event annotations for supporting
retrieving Values from ValueContainers contained within the event's
cause. GetValue is unique in that it can support not just the event's
cause's first ValueContainer, but also the preceeding ValueContainer
compatible parameter for the event listener. This means that an
annotated listener can be parameterized as follows:
```
@Listener
public void onDamageEntity(DamageEntityEvent event, @Getter("getTargetEntity") Zombie zombie, @GetValue("HEALTH") double
health) {

}
```

The advantages with these annotations are that they rely on using the
`class` container and the string being the direct field name. This means
that the field must be a `Key`, publicly accessible, and static. For
plugins, this will be easy to make use of for technically faster getters
of data than previously available, and supporting better event
filtering.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>