Data API v4: Migrating instance methods to DataProviders with Keys.
Essentially, because the old Data API required that all Keys were
semi-linked to serializable information, and with this refactor/rewrite,
we can take advantage that many of the instance methods that existed on
Entity based classes can now be refactored to use Keys which would be
backed by DataProviders. Of course, this means that most Entity
interfaces are now just filled with default methods and can be
eliminated, but for the sake of simplicity for plugin developers, we
retain the default "required" methods for getters, but may offer to
remove the "setter" methods that otherwise existed.

This commit is pending.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>