Properties exist in the WorldProperties.
The API was really messy in this area. We provided sugar methods simply
to allow people to get access to various properties of the
WorldProperties. The only issue is we only added like 5 out of 20+. If
we added all the sugars/forwarders to World/ServerWorld then this would
baloon out of control. WorldProperties exists for a very good reason, it
won't hurt anyone to have the extra method call.........

Signed-off-by: Chris Sanders <zidane@spongepowered.org>