Add overloads for Suppliers of CatalogTypes
Various methods in the API are designed to consume CatalogTypes
that are now statically used as Suppliers. Converting all methods
to use Suppliers is not ideal as plugins can be providing their
own CatalogTypes, so, we've added overloads that use both.

This commit also resolves the compilation issues preventing the
API from building previously.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>