It makes no sense to be able to build DisplaySlot.
There are only 3 distinctive DisplaySlot in Vanilla. The API was
violating the contract by allowing plugins to create them which violates
our CatalogType contract (it would use the same key each time). There is
no need for this when a simple setter for the color is all that is
necessary.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>