Add simple Placeholder system
This enables plugins to register their own PlaceholderParsers to the Sponge registry, enabling plugins to provide and access placeholders without needing a third party plugin.

Notably, this PR does NOT include a templating system. This is intentional, to allow plugins to provide their own systems as now.

Part of the ideas from #1891