Rename WorldLocation to Location and Location to ServerLocation
Updates usages throughout the API to using a ServerLocation in lieu of an abstract
Location for cases where the events/concepts require a Server to exist. Location is
still made semi usable by nature of being able to query for various things that may
exist on the client worlds, but restrict further logic to the ServerLocation case.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>