Fix several API design issues
- Add getServerLocation to Locatable that is used when the plugin dev is
sure it is on the server. Otherwise enjoy the crash. Minimizes
keystrokes
- Remove Player extending User. This was one of the biggest flaws of the
Player API in that the implementation had to hack around it and Users
passed around would always be Player objects. The original (and now
fulfilled) design was User would be created, when needed, to represent
the data of a Player. If the User is created and the Player is online
then the implementation can choose to glue it together or not. It is the
offline player data.
- Remove LocationCreator from World. API only allows creating
server-sided Locations. We allow Locations from the client to be
returned to disallow the client objects from being able to have logic
called off of them. This fixes that design issue originating from before
API 8.0.
- Downgrade WorldProperties to ServerWorld. ClientWorld has no concept
of the entirety of the level data.
- Separate out UserInventory to PlayerInventory per the User changes.

This cleans up a decent portion of the design problems once we decided
to, at a minimum, make Plugins aware that they may already being ran on
a Client.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>