Update HarvestEntityEvent for better exposure of what happens
The original event was somewhat "wishful thinking" in that we'd know
for every entity that is being handled, whereas the new implementation
allows for the event to be properly tracked whenever an Entity is being
"harvested" for drops and items, etc. The only reservation is that the
event is unaware of any further drops being performed as those will be
exposed separately in a related SpawnEntityEvent.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>