Refactor ChangeBlockEvent with All, Post, and Pre
Previous events were disjointed by a class definition of what possible
operations were being performed, regardless whether an operation was
performed in a specific order. This lead to either logically reading the
transactions in order from the Post event, and deducing changes, or
avoiding transaction ordering entirely. Now, with the new API, the
transaction itself lends the Operation as a signifier about what
operation the transaction is itself, without needing to know whether the
changes were done in a specific order or not, they're always in the
order the engine provides, with the added information for each transaction
to have the operation being performed to achieve the result of the final
block state.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>