Scoped registry framework.
As Minecraft continues to be more data-driven, the API must evolve and
serve as a guide to plugin developers. For game content, the engine is a
croassroads where 1.16.4 shows the start of game objects driven outside
code and clients receiving game content during connection.

This API hopes to future proof against such endeavours. Most importantly,
the API removes the concept of catalogs being singletons and treats
everything as references.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>