Holders should not be responsible for going through scopes.
That is the job of RegistryReference and most plugin devs will use those references regardless.

Might try and add some utility method somewhere, dunno where, for when you don't have a RegistryReference and want to go up scopes but we need to be careful that Vanilla may, in the future, decide to not check against indirect scopes should a direct scope not have a registry entry. Food for thought.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>