Restructure the World API into a template system.
This is by far an in-complete representation of the power of the Minecraft world gen system now available in 1.16.4 but it is enough to allow Sponge to load worlds via template files without enforcing other implementations to do so.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>

Various World API changes

- add `keepLoaded` to templates/properties.
- add `referenced` to RegistryReference
- ConstructWorldPropertiesEvent is always server-only

...and other minor changes.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>

Define long term API method strategy (or start too)

- Add "defaultWorld" to WorldManager.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>

Add a configuration system for complex biome providers.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>