Desupply the API
Most of the supplier methods existed when types were going to be fed by
explicit suppliers (and not by lookups that references do right now).
Whereas references are suppliers, that was only done for passing to
other java constructs easier. The plugin dev needs to always call get
themselves to terminate the callstack at the call site.

It also doesn't help that the only consistent thing regarding suppliers
were we were inconsistent which makes for a messier API. This cleans
this up.

Lastly, ItemStack and BlockSnapshot both gain Factory objects to return
their empty references. Should NEVER generate a builder instance just to
do nothing with it...

Signed-off-by: Chris Sanders <zidane@spongepowered.org>