Refactor statistics to be implementable
- Remove Criterion from being in the registry as no registration process
exists to properly register new criterion. A factory is provided to
acquired Vanilla types
- Correct the registry for STATISTIC to be the keys (So a Registry of
Keys by Keys). This now follows Vanilla spec
- Expose the StatisticFormatter to the API so that mod formatters on
SpongeForge are properly provided to plugins. This is not complete, an
implementation issue on Sponge needs to be addressed to allow this to
work completely
- Correctly dictate that a Statistic IS a Criterion which now bridges
over into Scoreboard code correctly
- Introduce new API onto a StatisticCategory to allow adding new
statistics to a category
- Fixup javadocs in a few spots (this is not complete)

Signed-off-by: Chris Sanders <zidane@spongepowered.org>